✨ アプリケーション終了時に子プロセスをkill

+++ src-electron/util/subprocess.ts
@@ -1,23 +1,30 @@
 import { Failable } from 'src-electron/api/failable';
 import * as child_process from 'child_process';
 import { utilLoggers } from './logger';
+import { sleep } from './sleep';
+import { onQuit } from '../lifecycle/lifecycle';
 
 const loggers = utilLoggers.subprocess;
 
 export type ChildProcessPromise = Promise<Failable<undefined>> & {
-  kill(signal?: number | NodeJS.Signals | undefined): boolean;
+  finished(): boolean;
+  kill(signal?: number | NodeJS.Signals | undefined): Promise<void>;
   write(msg: string): Promise<void>;
 };
 
 function promissifyProcess(
   process: child_process.ChildProcess,
   processpath: string,
-  args: string[]
+  args: string[],
+  beforeKill: (child: ChildProcessPromise) => void | Promise<void> = () => {},
+  beforeKillTimeout = 1000
 ) {
   const logger = loggers.promissifyProcess({
     command: processpath + ' ' + args.join(' '),
   });
   logger.start();
+
+  let isFinished = false;
 
   function onExit(code: number | null) {
     if (code === 0 || code === null) return undefined;
@@ -34,19 +41,36 @@ function promissifyProcess(
   ) => void = (resolve) => {
     process.on('exit', (code) => {
       logger.success(code);
+      isFinished = true;
+      // プロセスkillの購読を解除
+      dispatch();
       resolve(onExit(code));
     });
     process.on('error', (err) => {
       logger.fail(err);
+      isFinished = true;
+      // プロセスkillの購読を解除
+      dispatch();
       resolve(err);
     });
   };
 
   const promise = new Promise(executor) as ChildProcessPromise;
 
-  const kill = (signal?: number | NodeJS.Signals | undefined): boolean => {
-    return process.kill(signal);
+  promise.finished = () => isFinished;
+
+  const kill = async (
+    signal?: number | NodeJS.Signals | undefined
+  ): Promise<void> => {
+    if (process.exitCode === null) {
+      // killの前処理
+      await Promise.any([beforeKill(promise), sleep(beforeKillTimeout)]);
+      process.kill(signal);
+    }
   };
+
+  // アプリケーション終了時にプロセスをkill
+  const dispatch = onQuit(() => kill(), true);
 
   const write = (msg: string) =>
     new Promise<void>((resolve) =>
@@ -64,7 +88,9 @@ export const interactiveProcess = (
   onout: ((chunk: string) => void) | undefined,
   onerr: ((chunk: string) => void) | undefined,
   cwd: string | undefined = undefined,
-  shell = false
+  shell = false,
+  beforeKill: (child: ChildProcessPromise) => void | Promise<void> = () => {},
+  beforeKillTimeout = 1000
 ): ChildProcessPromise => {
   const child = child_process.spawn(process, args, {
     cwd,
@@ -82,7 +108,15 @@ export const interactiveProcess = (
     child.stderr?.on('data', onerr);
   }
 
-  return promissifyProcess(child, process, args);
+  const result = promissifyProcess(
+    child,
+    process,
+    args,
+    beforeKill,
+    beforeKillTimeout
+  );
+
+  return result;
 };
 
 export function execProcess(
